Initial compatibility fix
* Changes to queries in the helpers and models so that they work with the
mapascoletivos schema. For example, the incident table doesn't have the
location_id column.

* The relationship between tables has been wired into the models. In this way,
we get to lazily load tuples (table rows) as ORM objects. This is was necessary
because of the social features built into mapascoletivos

* Modified the JSON controller to generate the location markers for the main map,
report detail map and user profile map (which shows all the geo-entries that the
user has contributed to the system)