Implement user signup
* When users sign up, they are sent an email that contains the link that
allows them to proceed with creating an account. The link contains
and activation token which expires after 24 hours. The activation tokens
are stored in the auth_tokens table

* Fixed the selection of features on the map. These are now attached to the
SelectFeature control; previously they were attached to the layer. Also, the
popups are now created using DOM methods as opposed to stringing HTML

* Do not redraw the map layers when the zoom on the main map changes. This
is because clustering is off by default; TODO: Find a better way of switching
this on/off