Only show locations from public maps
* The main map now shows locations from public maps only.
The same restriction has also been cascaded to he reports
listing page so visitors can only browse the public maps.
Also, when viewing a map, the application checks
if the currently logged in user has permission to view the map.
Only the owner of the map, or a user with the "superadmin" role
can view a private map. Closes #10

* Fix logout link in the admin console

* Removed redundant view files not moved around during the
upgrade to 2.5