Create location using "Find location" address bar
* Locations can now be created using the "find location"
address bar - refs #7. Waiting for @julianamori to verify
that the feature works before closing. 

* Refactored Layer_Model::is_valid_layer to simply
check for existence of a layer record with the specified
ID

* Creating new locations using the point tool has now been
resolved. The add/edit location dialog was not opening because
the incident_legend_id property of the Location model (Backbone)
was undefined. Closes #12

* Added Layer_Model::get_visible_layers - a helper method
for retrieving all the visible layers